# Use the Regex source generator in the AlphaRouteConstraint.