--- a/src/Http/Routing/src/Constraints/AlphaRouteConstraint.cs
+++ b/src/Http/Routing/src/Constraints/AlphaRouteConstraint.cs
@@ -1,17 +1,22 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Text.RegularExpressions;
+
 namespace Microsoft.AspNetCore.Routing.Constraints;
 
 /// <summary>
 /// Constrains a route parameter to contain only lowercase or uppercase letters A through Z in the English alphabet.
 /// </summary>
-public class AlphaRouteConstraint : RegexRouteConstraint
+public partial class AlphaRouteConstraint : RegexRouteConstraint
 {
     /// <summary>
     /// Initializes a new instance of the <see cref="AlphaRouteConstraint" /> class.
     /// </summary>
-    public AlphaRouteConstraint() : base(@"^[a-z]*$")
+    public AlphaRouteConstraint() : base(GetAlphaRouteRegex())
     {
     }
+
+    [GeneratedRegex(@"^[A-Za-z]*$", RegexOptions.CultureInvariant)]
+    private static partial Regex GetAlphaRouteRegex();
 }