--- a/src/Http/Routing/src/Constraints/RegexRouteConstraint.cs
+++ b/src/Http/Routing/src/Constraints/RegexRouteConstraint.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Text.RegularExpressions;
 using Microsoft.AspNetCore.Http;
@@ -33,7 +34,7 @@ public class RegexRouteConstraint : IRouteConstraint, IParameterLiteralNodeMatch
     /// Constructor for a <see cref="RegexRouteConstraint"/> given a <paramref name="regexPattern"/>.
     /// </summary>
     /// <param name="regexPattern">A string containing the regex pattern.</param>
-    public RegexRouteConstraint(string regexPattern)
+    public RegexRouteConstraint([StringSyntax(StringSyntaxAttribute.Regex)] string regexPattern)
     {
         if (regexPattern == null)
         {