The Regex source generator is faster because it doesn't need to interpret the regex. It does add slightly more code to the compiled app, but the tradeoff seems worth it.

I switched the pattern from `a-z` with IgnoreCase to `A-Za-z` because the source generator generates better code with that pattern.

The results of the changed Benchmark shows a 9-10% faster when using an `alpha` constraint:

### Before

|          Method |     Mean |     Error |    StdDev |      Op/s | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------:|----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|      TreeRouter | 1.362 us | 0.0107 us | 0.0100 us | 734,023.0 |  1.00 | 0.0134 |     - |     - |      1 KB |
| EndpointRouting | 1.763 us | 0.0107 us | 0.0100 us | 567,280.4 |  1.29 | 0.0134 |     - |     - |      1 KB |


### After

|          Method |     Mean |     Error |    StdDev |      Op/s | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------:|----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|      TreeRouter | 1.230 us | 0.0087 us | 0.0082 us | 813,331.4 |  1.00 | 0.0134 |     - |     - |      1 KB |
| EndpointRouting | 1.615 us | 0.0068 us | 0.0063 us | 619,255.2 |  1.31 | 0.0153 |     - |     - |      1 KB |

### Size

Publishing a Linux NativeAOT app shows about a 4 KB increase in code size with this change:

**Before**: 27.1 MB (28,483,016 bytes)
**After** : 27.1 MB (28,487,464 bytes)