@gfoidl 2022-10-28
Note for further optimization potential:
Instead of relying on Regex for this, the `Match` could be expressed in terms of https://github.com/dotnet/runtime/issues/76106 or https://github.com/dotnet/runtime/issues/68328, thus avoiding Regex's overhead (and allocations).