@danmoseley 2022-10-28
> I switched the pattern from a-z with IgnoreCase to A-Za-z because the source generator generates better code with that pattern.

Is there a regex issue for that? @stephentoub