@stephentoub 2022-10-28
>> I switched the pattern from a-z with IgnoreCase to A-Za-z because the source generator generates better code with that pattern.

> Is there a regex issue for that? @stephentoub (This is invariant culture)

There's no "issue"... it's by design: [a-z] (IgnoreCase) and [A-Za-z] are functionally different... the former recognizes the Kelvin sign as being case-equivalent to k/K.