@eerhardt 2022-10-28
> Instead of relying on Regex for this,

The public inheritance here is a problem.

AlphaRouteConstraint is a public class that inherits from RegexRouteConstraint. 

1. RegexRouteConstraint's constructors take a Regex or a string. I would need to make a new internal ctor just for this class.
2. RegexRouteConstraint's implementation of `IRouteConstraint.Match` is not virtual. In order to make this work without any breaking changes, I would need to explicitly implement `IRouteConstraint.Match`, and you would get different behavior depending on how you cast the object.