@eerhardt 2022-10-28
>  the former recognizes the Kelvin sign as being case-equivalent to k/K.

Sounds like that was an existing bug, given its documentation: https://learn.microsoft.com/en-us/dotnet/api/system.web.mvc.routing.constraints.alpharouteconstraint

> Constrains a route parameter to contain only lowercase or uppercase letters A through Z in the English alphabet.

for what it is worth:

https://xkcd.com/
https://xKcd.com/
https://xKcd.com/

All take me to the same place.