@danmoseley 2022-10-28
> With more work and analysis we could optimize the latter to instead be:
char.IsAsciiLetter(ch = slice[iteration]) || ch == '\u212A'

On the face of it that seems like that would be a worthwhile improvement for a fairly common looking pattern?